migrating dynamo conditional failure function to shared helper
